Move SpasAdapter to new package.
