+++ client/src/main/java/com/alibaba/nacos/client/auth/ram/RamClientAuthServiceImpl.java
@@ -24,7 +24,7 @@ import com.alibaba.nacos.client.auth.ram.injector.ConfigResourceInjector;
 import com.alibaba.nacos.client.auth.ram.injector.NamingResourceInjector;
 import com.alibaba.nacos.client.auth.spi.AbstractClientAuthService;
 import com.alibaba.nacos.client.auth.spi.RequestResource;
-import com.alibaba.nacos.client.config.impl.SpasAdapter;
+import com.alibaba.nacos.client.auth.ram.utils.SpasAdapter;
 import com.alibaba.nacos.client.auth.ram.identify.StsConfig;
 import com.alibaba.nacos.common.utils.StringUtils;
 import org.slf4j.Logger;
