+++ client/src/main/java/com/alibaba/nacos/client/auth/ram/utils/SpasAdapter.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 1999-2018 Alibaba Group Holding Ltd.
+ * Copyright 1999-2021 Alibaba Group Holding Ltd.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.alibaba.nacos.client.config.impl;
+package com.alibaba.nacos.client.auth.ram.utils;
 
 import com.alibaba.nacos.api.common.Constants;
 import com.alibaba.nacos.client.auth.ram.identify.CredentialService;
